refactor: refactor the script to make it simplier
